new header, layout, and blocks route
