+++ apps/www/app/components/table-of-contents.tsx
@@ -1,5 +1,4 @@
 import { cx } from "@ngrok/mantle/cx";
-import { ListNumbersIcon } from "@phosphor-icons/react/ListNumbers";
 import { useCallback, useEffect, useRef, useState, type RefObject } from "react";
 import { createPortal } from "react-dom";
 import { Link, useLocation } from "react-router";
@@ -183,12 +182,9 @@ function TableOfContents({ contentRef }: { contentRef: RefObject<HTMLDivElement 
 	}
 
 	return createPortal(
-		<nav aria-label="Table of contents" className="sticky top-0 max-h-screen w-40">
-			<div className="scrollbar h-full overflow-y-auto py-4">
-				<p className="text-strong mb-3 flex items-center gap-1.5 text-xs font-medium uppercase tracking-wide">
-					<ListNumbersIcon className="size-4" />
-					On this page
-				</p>
+		<nav aria-label="Table of contents" className="sticky top-20 max-h-[calc(100vh-5rem)] w-40">
+			<div className="scrollbar h-full overflow-y-auto pb-4">
+				<p className="mb-3 text-xs font-medium uppercase tracking-wider font-mono">On this page</p>
 				<ol className="space-y-1 border-l border-gray-300">
 					{entries.map((entry) => (
 						<li key={entry.id}>
