+++ apps/www/app/root.tsx
@@ -27,12 +27,13 @@ import {
 	useRouteLoaderData,
 } from "react-router";
 import type { Route } from "./+types/root";
-import { Layout as WwwLayout } from "./components/layout";
 import { NavigationProvider } from "./components/navigation-context";
 import { useNonce } from "./components/nonce";
 import "./global.css";
 import { canonicalDomain, canonicalHref } from "./utilities/canonical-origin";
+import { parseMantleVersion } from "./utilities/mantle-version.server";
 import invariant from "tiny-invariant";
+import { MantleVersionProvider } from "./components/mantle-version-provider";
 
 const themeUrls = mantleStyleSheetUrls({
 	darkCssUrl,
@@ -101,12 +102,15 @@ export const loader = async ({ request }: Route.LoaderArgs) => {
 	const nodeEnv = process.env.NODE_ENV ?? "development";
 
 	return {
-		currentVersion: packageJson.default.version,
+		/**
+		 * The current version of mantle from the package.json
+		 */
+		currentMantleVersion: parseMantleVersion(packageJson.default.version),
 		commitSha,
 		deploymentId,
 		renderReactQueryDevtools: nodeEnv !== "production",
 		ssrCookie: extractThemeCookie(request.headers.get("Cookie")),
-	};
+	} as const;
 };
 
 export function shouldRevalidate(_: ShouldRevalidateFunctionArgs) {
@@ -145,8 +149,8 @@ export function Layout({ children }: PropsWithChildren) {
 		window.toggleReactQueryDevtools = () => setShowReactQueryDevtools((previous) => !previous);
 	}, []);
 
-	const { currentVersion } = loaderData ?? {};
-	invariant(currentVersion, "current version should be defined");
+	const { currentMantleVersion } = loaderData ?? {};
+	invariant(currentMantleVersion, "current version should be defined");
 
 	return (
 		<html {...initialHtmlThemeProps} lang="en-US" dir="ltr" suppressHydrationWarning>
@@ -170,7 +174,7 @@ export function Layout({ children }: PropsWithChildren) {
 			</head>
 			<body
 				className={cx(
-					"bg-base h-full min-h-full overflow-y-scroll scrollbar isolate relative",
+					"bg-card h-full min-h-full overflow-y-scroll scrollbar isolate relative",
 					scrollBehavior === "smooth" && "scroll-smooth",
 				)}
 			>
@@ -183,9 +187,9 @@ export function Layout({ children }: PropsWithChildren) {
 									<ReactQueryDevtoolsLazy />
 								</Suspense>
 							)}
-							<NavigationProvider>
-								<WwwLayout currentVersion={currentVersion}>{children}</WwwLayout>
-							</NavigationProvider>
+							<MantleVersionProvider mantleVersion={currentMantleVersion}>
+								<NavigationProvider>{children}</NavigationProvider>
+							</MantleVersionProvider>
 						</QueryClientProvider>
 					</TooltipProvider>
 				</ThemeProvider>
